feat: support raw Vue SSR
