+++ __tests__/ssr/app/entry-server.ts
@@ -1,8 +1,16 @@
+import Vue from 'vue'
 import { createApp } from './main'
+import { PiniaSsr, getRootState } from '../../../src'
+
+// simulate server-side environment
+// @ts-ignore
+delete global.window
+
+Vue.use(PiniaSsr)
 
 export default function(context: any) {
   return new Promise(resolve => {
-    const { app, store } = createApp()
+    const { app } = createApp()
 
     // This `rendered` hook is called when the app has finished rendering
     context.rendered = () => {
@@ -11,7 +19,7 @@ export default function(context: any) {
       // When we attach the state to the context, and the `template` option
       // is used for the renderer, the state will automatically be
       // serialized and injected into the HTML as `window.__INITIAL_STATE__`.
-      context.state = store.state
+      context.state = getRootState(context.req)
     }
 
     resolve(app)
