+++ src/index.ts
@@ -1,3 +1,4 @@
 export { createStore } from './store'
 export { setActiveReq, setStateProvider, getRootState } from './rootStore'
 export { StateTree, StoreGetter, Store } from './types'
+export { PiniaSsr } from './ssrPlugin'
